pyproject.toml
[project]
name = "nonebot-plugin-example"
dynamic = ["version"]
description = "Template plugin project"
authors = [{ name = "Murasame", email = "congyu@sbhfy.cn" }]
dependencies = ["nonebot2>=2.4.1", "nonebot-plugin-alconna>=0.54.2"]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/LeiSureLyYrsc/nonebot-plugin-template"

[dependency-groups]
dev = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_example/__init__.py"
